Rename project to CppHTTP, fix a bug in std::format
