+++ CppHTTP.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29728.190
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "WS++", "WS++\WS++.vcxproj", "{C897069C-1CBC-4B6D-8304-0DF101591CA6}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CppHTTP", "CppHTTP\CppHTTP.vcxproj", "{C897069C-1CBC-4B6D-8304-0DF101591CA6}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "WSTest", "WSTest\WSTest.vcxproj", "{6CA48AF2-A658-4773-97D3-AB19E36406CF}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Tests", "Tests\Tests.vcxproj", "{6CA48AF2-A658-4773-97D3-AB19E36406CF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
