+++ CppHTTP/Common.h
