+++ CppHTTP/Common/Standard/Errors.h
