+++ CppHTTP/Common/Standard/Strings.h
