+++ CppHTTP/Common/Standard/Types.h
