+++ CppHTTP/HTTP/Errors.h
