+++ CppHTTP/Sockets/AddresInfo.cpp
