+++ CppHTTP/Sockets/AddresInfo.h
