+++ CppHTTP/Sockets/Socket.cpp
