+++ CppHTTP/Sockets/Socket.h
