+++ CppHTTP/Sockets/SocketIO.h
