+++ CppHTTP/Sockets/WSA.cpp
